refactored textinput
